# Move object_hook outside loads method so class can be extend and reused