Wanted to be able to extend the session serializer class without duplicating all the code for object hook - as with the _tag function above.