@ThiefMaster 2015-05-25
Wouldn't it be better to make it a static method of `TaggedJSONSerializer`? That way you can simply subclass it instead of having to rewrite the dumps/loads methods of that class or monkeypatch.

In the same go, it'd be nice to make those two methods static, too, and use the class itself instead of an instance of it. Those methods never access `self` anyway...