@ThiefMaster 2015-06-20
While you're at it, maybe #1511 could be interesting to fix at the same time.