@ThiefMaster 2015-06-20
I'd move the untagging code into a method, too. It's a bit pointless to be able to override it for only one direction