@ThiefMaster 2015-06-20
This is what I'd do btw (started working on it earlier today before I remembered that there's already your PR).

```python
class TaggedJSONSerializer(object):
    """A customized JSON serializer that supports a few extra types that
    we take for granted when serializing (tuples, markup objects, datetime).
    """

    def dumps(self, value):
        return json.dumps(self._tag(value), separators=(',', ':'))

    def loads(self, value):
        def object_hook(obj):
            if len(obj) != 1:
                return obj
            the_key, the_value = next(iteritems(obj))
            rv = self._untag(the_key, the_value)
            return obj if rv is None else rv
        return json.loads(value, object_hook=object_hook)

    def _untag(self, key, value):
        if key == ' t':
            return tuple(value)
        elif key == ' u':
            return uuid.UUID(value)
        elif key == ' b':
            return b64decode(value)
        elif key == ' m':
            return Markup(value)
        elif key == ' d':
            return parse_date(value)
        return None

    def _tag(self, value):
        if isinstance(value, tuple):
            return {' t': [self._tag(x) for x in value]}
        elif isinstance(value, uuid.UUID):
            return {' u': value.hex}
        elif isinstance(value, bytes):
            return {' b': b64encode(value).decode('ascii')}
        elif callable(getattr(value, '__html__', None)):
            return {' m': text_type(value.__html__())}
        elif isinstance(value, list):
            return [self._tag(x) for x in value]
        elif isinstance(value, datetime):
            return {' d': http_date(value)}
        elif isinstance(value, dict):
            return dict((k, self._tag(v)) for k, v in iteritems(value))
        elif isinstance(value, str):
            try:
                return text_type(value)
            except UnicodeError:
                from flask.debughelpers import UnexpectedUnicodeError
                raise UnexpectedUnicodeError(u'A byte string with '
                    u'non-ASCII data was passed to the session system '
                    u'which can only store unicode strings.  Consider '
                    u'base64 encoding your string (String was %r)' % value)
        return value
```