@s-block 2015-07-10
I have added a fix for issue #1511 as well as making the class extendible, could someone have a look and see what you think?