@ThiefMaster 2015-07-10
Looks quite complex (don't have time to thoroughly review it right now). Also, it adds a list of valid tags at yet another place, making it harder to extend... maybe it'd be worth refactoring the whole thing to change the way the conversions are defined?

```
conversions = {
    't': {'check': lambda value: isinstance(value, tuple),
          'tag': lambda value: [self._tag(x) for x in value]},
          'untag': lambda value: tuple(value)},
    ...
}
```

or something similar