@s-block 2015-07-17
So I have done something along the lines of your suggestion, conversions couldn't be a dict as it needs to retain order of checks, OrderedDict wasn't in python until 2.7 so tests would fail so I have made a list.

Also some tag/untag's couldn't be lambdas so had to declare self.conversions in __init__ to access self._custom_tag/untag methods

Any thoughts? Looks like more code but will be much easier to extend:

```python
def __init__:
    super(TaggedJSONSerializer, self).__init__()
    self.conversions.append(...)
```