@RonnyPfannschmidt 2016-11-22
did anyone check the performance on this? compared to current methods and #2093