@s-block 2017-02-23
I did a quick performance test on the session and the proposed code is slightly slower but not too different.

The test code was:

```python
import cProfile
from datetime import datetime
import uuid

import flask
from flask.sessions import TaggedJSONSerializer


def test_session_performance():
    now = datetime.utcnow().replace(microsecond=0)
    the_uuid = uuid.uuid4()

    data = [
        (1, 2, 3),
        the_uuid,
        b'\xff',
        flask.Markup('Hello!'),
        now,
    ]

    cycles = 10000

    session = TaggedJSONSerializer()

    for n, d in enumerate(data * cycles):
        dumped = session.dumps(d)
        session.loads(dumped)


if __name__ == '__main__':
    cProfile.run('test_session_performance()')

```

Latest flask master branch:
4.319 seconds
4.409 seconds
4.455 seconds

This branch:
4.891 seconds
4.913 seconds
4.957 seconds