@s-block 2017-04-26
Is this OK to be merged? It would be good to get this functionality in.