@davidism 2017-04-26
I'm working through the PR backlog right now, this is on the list. Would you mind enabling the "allow edits from maintainers option" in the sidebar?