@kennethreitz 2017-05-25
resolved merge conflict. waiting for tests to pass