@kennethreitz 2017-05-25
omg this test matrix is obscene