@davidism 2017-06-01
I was reviewing this and decided it's not extensible enough. I'm rewriting it, but on top of your commits so you still get credit, since most of the effort was yours. Added some review comments, but they're already fixed on my local branch.

New idea is to have a `JSONTag` interface, and give `TaggedJSONSerializer` a `register` method to add new tags.