@davidism 2017-06-01
Hmm, maybe this should all be moved to itsdangerous. It's not actually Flask specific.