@davidism 2017-06-02
The only place it's being used is passed to itsdangerous. Eh, I could go either way, probably not worth the trouble.